css and dropdown update
Finished the dropdown for the most part. It is functional, but have to make it responsive for smaller screen sizes. Delete the nav on the footer and replace with a up arrow. Added second round of images to programming page need to do the same for the writing page. Added download button for JS certificate and resume.